Broke out all the operational functions into operations.go
Added runtime configurations to .run

Renamed train package to training

More general refactorings
